+++ NEWS.md
@@ -1,3 +1,8 @@
-# tr808r 0.1.0
+# tr808r 0.1.1  2022-05-14
+
+* Fixed pattern length bug when loading pattern from a tweet
+
+
+# tr808r 0.1.0  2022-05-13
 
 * Initial release